refining the examples a little bit
