+++ demo/demo1.py
@@ -1,4 +1,4 @@
-from dryvr_plus_plus.example.example_agent.car_agent import CarAgent
+from dryvr_plus_plus.example.example_agent.car_agent import CarAgent, NPCAgent
 from dryvr_plus_plus.scene_verifier.scenario.scenario import Scenario
 from dryvr_plus_plus.example.example_map.simple_map2 import SimpleMap2, SimpleMap3, SimpleMap5, SimpleMap6
 from dryvr_plus_plus.plotter.plotter2D import *
@@ -34,33 +34,35 @@ if __name__ == "__main__":
     input_code_name = 'example_controller1.py'
     scenario = Scenario()
 
-    car = CarAgent('car1', file_name=input_code_name)
+    car = NPCAgent('car1')
     scenario.add_agent(car)
     car = CarAgent('car2', file_name=input_code_name)
     scenario.add_agent(car)
-    tmp_map = SimpleMap6()
+    tmp_map = SimpleMap3()
     scenario.set_map(tmp_map)
     scenario.set_sensor(FakeSensor2())
     scenario.set_init(
         [
-            [[0, -0.2, 0, 1.0],[0.1, 0.2, 0, 1.0]],
-            [[10, 0, 0, 0.5],[10, 0, 0, 0.5]], 
+            [[10.0, 0, 0, 0.5],[10.0, 0, 0, 0.5]], 
+            [[5.0, -0.2, 0, 2.0],[6.0, 0.2, 0, 3.0]],
         ],
         [
             (VehicleMode.Normal, LaneMode.Lane1),
             (VehicleMode.Normal, LaneMode.Lane1),
         ]
     )
-    # res_list = scenario.simulate_multi(40,1)
-    traces = scenario.verify(40)
+    res_list = scenario.simulate_multi(10,1)
+    # traces = scenario.verify(10)
 
     fig = plt.figure(2)
-    fig = plot_map(tmp_map, 'g', fig)
-    fig = plot_reachtube_tree(traces, 'car1', 1, [2], 'b', fig)
-    fig = plot_reachtube_tree(traces, 'car2', 1, [2], 'r', fig)
-    # for traces in res_list:
-    #     generate_simulation_anime(traces, tmp_map)
-    #     # fig,x_lim,y_lim = plot_simulation_tree(traces, 'car1', 1, [2], 'b', fig,x_lim,y_lim)
-    #     # fig,x_lim,y_lim = plot_simulation_tree(traces, 'car2', 1, [2], 'r', fig,x_lim,y_lim)
+    # fig = plot_map(tmp_map, 'g', fig)
+    # fig = plot_reachtube_tree(traces, 'car1', 0, [1], 'b', fig, (1000,-1000), (1000,-1000))
+    # fig = plot_reachtube_tree(traces, 'car2', 0, [1], 'r', fig)
+    for traces in res_list:
+        # fig = plot_simulation_tree(traces, 'car1', 0, [1], 'b', fig, (1000,-1000), (1000,-1000))
+        # fig = plot_simulation_tree(traces, 'car2', 0, [1], 'r', fig)
+        generate_simulation_anime(traces, tmp_map, fig)
+
+
 
     plt.show()
